Fix warning in ball_userfieldsschama.h
Warning on converting a 'size_t' to an 'int', which could
be a significant narrowing on a 64-bit platform.  That said,
2 billion fields in a schema should be more than enough for
much longer than the immediate future, so the easiest fix is
to cast to 'int', rather than change the return type.

Considered adding a 'BSLS_ASSERT_SAFE' that the value was
less than 'numeric_limits<int>::max()', but did not want to
add an extra header dependency in this header.  Would have
done so if the method were out-of-line in the .cpp file.

+++ groups/bal/ball/ball_userfieldsschema.h
@@ -371,7 +371,7 @@ ball::UserFieldType::Enum UserFieldsSchema::type(int index) const
 inline
 int UserFieldsSchema::length() const
 {
-    return d_names.size();
+    return static_cast<int>(d_names.size());
 }
 
 }  // close package namespace
